Add GeneralStudy module and helper functions.
Implemented a GeneralStudy module to describe and identify the study performed on the patient. It provides the structure and the attributes needed to represent a study. Helper functions were added to aid in the parsing of DICOM objects. The module 'general_study' was exposed in 'modules.rs'. Also, 'helper.rs' was created to include the helper function 'element_opt_to_str' that reads the string value of an optional DICOM tag. These changes allow facilitating more robust handling and manipulation of DICOM studies.